- doing rendering

+++ src/main/java/net/smackem/jobotwar/runtime/Constants.java
@@ -7,8 +7,8 @@ public class Constants {
 
     public static int MAX_SHOT = 10_000;
     public static double MAX_RANGE = 10_000;
-    public static double MAX_BOARD_WIDTH  = 5_000;
-    public static double MAX_BOARD_HEIGHT  = 5_000;
+    public static int MAX_BOARD_WIDTH  = 5_000;
+    public static int MAX_BOARD_HEIGHT  = 5_000;
     public static double ANGLE_PRECISION = 0.0001;
     public static int MAX_HEALTH = 100;
     public static double MAX_ROBOT_ACCELERATION = 10;
